requirements.txt
pyyaml
fastapi

